uploaded base files
All the files created offline before the creation of this repo was uploaded, note that this is just the front end of the web application, the web app on completion will have both front end and a back end, so it can only be hosted on the local server. Enjoy the front end on the website until you can.